ai_service/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# OpenAI
openai==1.7.1

# ML & NLP
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4

# Document processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
tenacity==8.2.3
